+++ src/app/utils.py
@@ -1,0 +1,47 @@
+"""Здесь хранятся дополнительные переменные и функции для работы бота."""
+
+
+#Сообщение, которым бот отвечает на мета-вопросы в чате группы.
+GROUP_MESSAGE = """
+Ребят, пожалуйста, не задавайте <a href='https://nometa.xyz/ru.html'>мета-вопросы</a> в чате!
+– вопросы, которые подразумевают другие вопросы.
+
+Примеры мета-вопросов:
+- Можно задать вопрос про X?
+- Здесь есть, кто проходил курс Y?
+- Ребят, а кто работал с Z?
+
+Они тратят время! И ваше, и других людей, которые пытаются вам помочь :)
+"""
+
+patterns_meta_questions = [
+    'здесь кто-нибудь работал с',
+    'кто-нибудь разбирается в',
+    'можно задать вопрос про',
+    'кто проходил курс',
+    'кто шарит за',
+    'кто юзал',
+    'можно задать вопрос',
+]
+
+
+def is_meta_question(message: str) -> bool:
+    """
+    Функция проверяет, является ли сообщение мета-вопросом.
+
+    Parameters
+    ----------
+    message : str
+        Сообщенние от пользователя.
+
+    Returns
+    -------
+    bool
+        True, если сообщение - мета вопрос.
+        False, если сообщение - обычный вопрос.
+    """
+    message = message.lower()
+    for meta_question in patterns_meta_questions:
+        if meta_question in message:
+            return True
+    return False
